First stab at bringing this up to WebVR spec
It looks like getVRDevices is an older method; the newer one is now
getVRDisplays. In my version of experimental Chromium, getVRDevices
throws an unknown method error.

This diff gets us to the point where the extension loads on Youtube,
shows the Oculus button, and doesn't throw any errors; however, it just
hangs now (the oculus headset just says "Sorry, chrome.exe is taking a
while to load."). So this is a checkpoint commit until I figure out why
that's hanging.